+++ src/datatypes/external.ts
@@ -110,6 +110,7 @@ module Plywood {
     customAggregations?: CustomDruidAggregations;
     allowEternity?: boolean;
     allowSelectQueries?: boolean;
+    avoidSegmentMetadata?: boolean;
     exactResultsOnly?: boolean;
     context?: Lookup<any>;
     druidVersion?: string;
@@ -134,6 +135,7 @@ module Plywood {
     customAggregations?: CustomDruidAggregations;
     allowEternity?: boolean;
     allowSelectQueries?: boolean;
+    avoidSegmentMetadata?: boolean;
     exactResultsOnly?: boolean;
     context?: Lookup<any>;
     druidVersion?: string;